+++ PageParser.cpp
@@ -88,7 +88,7 @@ bool sendPageParser2 ( TcpSocket   *s ,
 		catch ( ... ) {
 			g_errno = ENOMEM;
 			log("PageParser: new(%i): %s", 
-			    sizeof(State8),mstrerror(g_errno));
+			    (int)sizeof(State8),mstrerror(g_errno));
 			return g_httpServer.sendErrorReply(s,500,
 						       mstrerror(g_errno));}
 		mnew ( st , sizeof(State8) , "PageParser" );
@@ -654,7 +654,7 @@ bool sendPageAnalyze ( TcpSocket *s , HttpRequest *r ) {
 	catch ( ... ) {
 		g_errno = ENOMEM;
 		log("PageParser: new(%i): %s", 
-		    sizeof(State8),mstrerror(g_errno));
+		    (int)sizeof(State8),mstrerror(g_errno));
 		return g_httpServer.sendErrorReply(s,500,
 						   mstrerror(g_errno));}
 	mnew ( st , sizeof(State8) , "PageParser" );
